+++ data/flowercare.js
@@ -59,6 +59,14 @@ function MyCallback(json) {
   
   document.querySelector("#loader").style.visibility = "hidden";
   document.querySelector("body").style.visibility = "visible";
+
+  var data = {};
+  data['cmd'] = {};
+  data['cmd']['action'] = "subscribe";
+  data['cmd']['subaction'] = "flowercare_data";
+  data["cmd"]["highlight"] = "true";
+  
+  global.requestData(data);
 }
 
 // ************************************************
@@ -90,15 +98,51 @@ export function ActivateDevice(id) {
 }
 
 // ************************************************
-function formatDate(obj) {
-  const intValue = parseInt(obj.innerText, 10);
+export function forceUpdate(button) {
+  var mac = button.closest('tr').querySelector('input[type="checkbox"][id$="active"]').getAttribute('data-mac');
+  var active = button.closest('tr').querySelector('input[type="checkbox"][id$="active"]').checked;
+  if (!active) {
+    global.setResponse(false,"Device is not active. Please activate it first to be able to update it.");
+    return;
+  }
+
+  var data = {};
+  data['cmd'] = {};
+  data['cmd']['action'] = "flowercare";
+  data['cmd']['subaction'] = "forceUpdate";
+  data['cmd']['item'] = mac;
+
+  global.requestData(data);
+}
+
+
+// ************************************************
+function formatDate(obj, TimeDiffInt) {
+  let intValue = 0;
+  if (TimeDiffInt) { intValue = TimeDiffInt; } else { intValue = parseInt(obj.innerText, 10); }
   if (intValue > 0) {
     var diff = esp_uptime - intValue;
     var hours = Math.floor(diff / 3600000);
     var minutes = Math.floor((diff % 3600000) / 60000);
     var seconds = Math.floor((diff % 60000) / 1000);
     obj.innerText = `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+    // sichere originalwert im Attribute data-timestamp
+    obj.setAttribute('data-timestamp', intValue);
   }
+}
+
+// ************************************************
+//rufe folgende funktion jede sekunde auf, um die Zeiten zu aktualisieren
+setInterval(updateTimestamps, 1000);
+function updateTimestamps() {
+  document.querySelectorAll('[id$="lastLiveDataUpdate"], [id$="lastBatteryUpdate"]').forEach(element => {
+    let TimeDiffInt = parseInt(element.getAttribute('data-timestamp'), 10);
+
+    if (TimeDiffInt) {
+      TimeDiffInt -= 1000; // ändere den Timestamp um 1 Sekunde
+      formatDate(element, TimeDiffInt); // aktualisiere die Anzeige
+    }
+  });
 }
 
 // ************************************************